+++ src/main/java/com/shawncurrie/techblogs/shared/dto/CompanyDTO.java
@@ -1,17 +1,15 @@
 package com.shawncurrie.techblogs.shared.dto;
 
-import javax.persistence.Column;
-
 public class CompanyDTO {
-    private long id;
+    private int id;
     private String name;
     private String url;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
